refactor: convert `apps/meteor/app/2fa/server/MethodInvocationOverride.js` to TypeScript
