+++ apps/meteor/app/api/server/ApiClass.ts
@@ -1050,7 +1050,7 @@ export class APIClass<TBasePath extends string = '', TOperations extends Record<
 					});
 
 					try {
-						const auth = await DDP._CurrentInvocation.withValue(invocation as any, async () => Meteor.callAsync('login', args));
+						const auth = await DDP._CurrentInvocation.withValue(invocation, async () => Meteor.callAsync('login', args));
 						const user = await Users.findOne(
 							{
 								_id: auth.id,
