+++ apps/meteor/definition/externals/meteor/ddp.d.ts
@@ -1,9 +1,8 @@
 import type { DDPCommon } from 'meteor/ddp-common';
+import type { Meteor } from 'meteor/meteor';
 
 declare module 'meteor/ddp' {
 	namespace DDP {
-		const _CurrentInvocation: {
-			withValue(invocation: DDPCommon.MethodInvocation, func: () => Promise<any>): Promise<any>;
-		};
+		const _CurrentInvocation: Meteor.EnvironmentVariable<DDPCommon.MethodInvocation>;
 	}
 }
